change type of radius and add ptr to args

+++ internal/handler/rent/dto/request.go
@@ -17,5 +17,5 @@ type UpdateRentRequest struct {
 type AvailableRentRequest struct {
 	Lat    float64 `json:"lat,omitempty" binding:"gte=-90,lte=90"`
 	Long   float64 `json:"long,omitempty" binding:"gte=-180,lte=180"`
-	Radius *uint64 `json:"radius,omitempty" binding:"omitempty,gt=0,lte=5000"`
+	Radius *uint16 `json:"radius,omitempty" binding:"omitempty,gt=0,lte=5000"`
 }
